feat: add themes switching
