+++ src/App.css
@@ -1,10 +1,5 @@
 #root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
   width: 100%;
-  font-size: larger;
 }
 
 .logo {
