+++ src/components/Button/Button.module.css
@@ -3,16 +3,20 @@
     border: 1px solid;
     cursor: pointer;
     font-size: 1em;
-    font-weight: 500;
+    line-height: 1.05rem;
     transition: border-color 0.25s;
     height: 44px;
     width: 110px;
     padding: 0;
     text-align: center;
     border-color: #0067b8;
-    color: #001526;
+    color: inherit;
 }
 
 button:hover {
     border-color: #646cff;
+  }
+
+  .dark {
+    background-color: transparent;
   }
