+++ src/components/Button/Button.tsx
@@ -1,5 +1,6 @@
-import { ReactNode } from 'react';
+import { ReactNode, useContext } from 'react';
 import styles from './Button.module.css';
+import { ThemeContext } from '../../contexts/ThemeContext';
 
 interface ButtonProps {
   className?: string;
@@ -9,12 +10,12 @@ interface ButtonProps {
   type?: 'submit' | 'reset' | 'button';
 }
 
-export function Button(props: ButtonProps) {
-  const { className, type, onClick, children } = props;
+export function Button({ className, type, onClick, children }: ButtonProps) {
+  const { isDarkTheme } = useContext(ThemeContext);
   return (
     <div className="button-wrapper">
       <button
-        className={`${className ? className + ' ' : ''} ${styles.button}`}
+        className={`${className ? className + ' ' : ''} ${styles.button} ${isDarkTheme ? styles.dark : styles.light}`}
         type={type}
         onClick={onClick}
       >
