+++ src/components/ListItem/ListItem.tsx
@@ -1,4 +1,6 @@
+import { useContext } from 'react';
 import styles from './ListItem.module.css';
+import { ThemeContext } from '../../contexts/ThemeContext';
 
 export interface ItemProps<T> {
   data: T;
@@ -8,8 +10,13 @@ export interface ItemProps<T> {
 }
 
 export function ListItem<T>({ data, renderDetails, id }: ItemProps<T>) {
+  const { isDarkTheme } = useContext(ThemeContext);
   return (
-    <li key={id} id={id} className={styles.searchResultsItem}>
+    <li
+      key={id}
+      id={id}
+      className={`${styles.searchResultsItem} ${isDarkTheme ? styles.dark : styles.light}`}
+    >
       {renderDetails(data)}
     </li>
   );
