+++ src/components/Pagination/Pagination.module.css
@@ -13,4 +13,5 @@
 .pagination {
   display: flex;
   justify-content: center;
+  flex-wrap: wrap;
 }
