+++ src/components/Root/Root.module.css
@@ -1,0 +1,23 @@
+.root {
+   
+    width: 100%;
+  
+}
+
+.container {
+    max-width: 1280px;
+    margin: 0 auto;
+    padding: 2rem 0;
+    text-align: center;
+    width: 100%;
+    font-size: larger;
+}
+
+.dark {
+    background-color: black;
+    color: #D7B7FD;
+}
+
+.light {
+    background-color: white;
+    }
