+++ src/components/SearchBar/SearchBar.module.css
@@ -2,15 +2,25 @@
   display: flex;
   justify-content: center;
   column-gap: 8px;
+  flex-wrap: wrap;
+  row-gap: 8px;
 }
 
 .searchInput {
   height: 38px;
   border-radius: 8px;
   font-size: larger;
+  background-color: transparent;
+  border: 1px solid;
+  color: inherit;
 }
 
 .searchButton {
   background-color: #0067b8;
   color: white;
+}
+
+.dark {
+  background-color: transparent;
+  color: inherit;
 }
