+++ src/contexts/ThemeContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+export const ThemeContext = createContext<{
+  isDarkTheme: boolean;
+  switchTheme: () => void;
+}>({ isDarkTheme: true, switchTheme: () => {} });
